Updated to allow for a dynamic list of files to be processed
Added ability to include a link to a file with a list of relative
file-paths in it to process in the analysis.
The separation is now simply done by checking if the parameter starts
with http if this is also the file name filter then it will try to
download it as a file first and will fail to do anything.